# docs/api/evalsmith-public.openapi.yaml
openapi: 3.1.0
info:
  title: EvalSmith Public API
  version: 0.1.0
  summary: Gateway-facing API surface for EvalSmith trial and private deployments
  description: >
    First-pass public contract for EvalSmith. This spec focuses on the core
    integration paths external users need for auth, datasets, traces,
    experiments, annotation, and monitoring.
servers:
  - url: http://127.0.0.1:8080
    description: Local gateway
...
paths:
...
  /api/v1/experiments:
    get:
      tags: [experiments]
      operationId: listExperiments
      summary: List experiments in the current project
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Experiment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
    post:
      tags: [experiments]
      operationId: createExperiment
      summary: Create and enqueue an experiment
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentCreateRequest"
      responses:
        "200":
          description: Experiment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /api/v1/experiments/{experiment_id}:
    get:
      tags: [experiments]
      operationId: getExperiment
      summary: Fetch an experiment and its summary
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/ExperimentIdPath"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Experiment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /api/v1/experiments/{experiment_id}/results:
    get:
      tags: [experiments]
      operationId: getExperimentResults
      summary: Fetch paginated experiment results
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/ExperimentIdPath"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, latency_ms, score]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: max_score
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Experiment results page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGenericEnvelope"
  /api/v1/experiments/target-preview:
    post:
      tags: [experiments]
      operationId: previewExperimentTarget
      summary: Preview how an endpoint target will be invoked for one example
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentTargetPreviewRequest"
      responses:
        "200":
          description: Target preview response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentTargetPreviewEnvelope"
  /api/v1/experiments/compare:
    post:
      tags: [experiments]
      operationId: compareExperiments
      summary: Compare two or more experiments
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperimentCompareRequest"
      responses:
        "200":
          description: Comparison result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      __REDACTED_SECRET__: API Key
    sessionCookie:
      type: apiKey
      in: cookie
      name: evalsmith_session
  parameters:
    ProjectIdHeader:
      name: X-Project-ID
      in: header
      required: true
      schema:
        type: string
      description: Active project identifier
...
    ExperimentIdPath:
      name: experiment_id
      in: path
      required: true
      schema:
        type: string
  schemas:
    EnvelopeBase:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        details:
          description: Optional error details
      required: [code, message]
    GenericEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data: {}
...
    PaginatedGenericEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: {}
                total:
                  type: integer
                page:
                  type: integer
                page_size:
                  type: integer
                total_pages:
                  type: integer
...
    ExperimentCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        dataset_id:
          type: string
        dataset_version:
          type: integer
          nullable: true
        split:
          type: string
        evaluator_ids:
          type: array
          minItems: 1
          items:
            type: string
        target_url:
          type: string
        target_method:
          type: string
          enum: [GET, POST, PUT, PATCH]
        target_headers:
          type: object
          additionalProperties:
            type: string
        target_body_template:
          type: string
        target_response_path:
          type: string
        target_timeout_ms:
          type: integer
        concurrency:
          type: integer
      required: [name, dataset_id, evaluator_ids, target_url]
    ExperimentPreviewExample:
      type: object
      properties:
        id:
          type: string
        inputs: {}
        expected_outputs: {}
        metadata:
          type: object
        split:
          type: string
      required: [inputs]
    ExperimentTargetPreviewRequest:
      type: object
      properties:
        target_url:
          type: string
        target_method:
          type: string
          enum: [GET, POST, PUT, PATCH]
        target_headers:
          type: object
          additionalProperties:
            type: string
        target_body_template:
          type: string
        target_response_path:
          type: string
        target_timeout_ms:
          type: integer
        example:
          $ref: "#/components/schemas/ExperimentPreviewExample"
      required: [target_url, example]
    ExperimentTargetPreviewEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                request_method:
                  type: string
                request_url:
                  type: string
                request_body: {}
                response_status_code:
                  type: integer
                response_path_used:
                  type: string
                  nullable: true
                latency_ms:
                  type: integer
                trace_id:
                  type: string
                  nullable: true
                output: {}
                raw_response: {}
    ExperimentCompareRequest:
      type: object
      properties:
        experiment_ids:
          type: array
          minItems: 2
          items:
            type: string
        baseline_experiment_id:
          type: string
          nullable: true
      required: [experiment_ids]
...